[debian] wait for property service by socket
If using getprop alone, it can fallback to /system/build.prop, thus not
ensuring the property is up. Some device (e.g. krillin) uses property
service to control wifi, thus this is required.

This stems from my ueventd-as-a-udevtrigger effort which prevents
double-coldbooting and fixes FP2's camera on Halium 7.1 port. However,
this means the start of udev doesn't guarantee the start of Android
container anymore. Thus, a better check is required.